Architecture refactored 2
